Update project structure:
New build, now copies the content of the julia folder (both includes and lib, even if actually the includes are not needed to be copied, might be removed). Also, it now boots with reference to the sys.dylib object copied over. It is a little fragile, since the path to the User Extensions folder is hardcoded in. Need to find a way to determine the folder where the Julia.scx is called dynamically.

Former-commit-id: 0c96434cce57b788e37f5504a843fb7104d175a3 [formerly e5bb79eb1567d6eae97550669af328fd906e34a4]
Former-commit-id: ad64c5b39f531dc45fc4c6a3b9af3d21895865ce