Fixed clicks. they were caused by the calls to popen
Calls to popen and pclose are blocking, that's why I would still block the audio thread if I was calling get_julia_dir() in the include function. I simply now have a global variable with the location and just append the name of the julia file to load. Anyway, get_julia_dir() is needed to run just once at boot.

Former-commit-id: 6d33b38b1f7be3d4f5df6549b3decfe6d2055dae [formerly 56271486e90fe96aaf59812fc81a4e15e65815b9]
Former-commit-id: 9f87962823944911239c67198664dedca1e089ff